Handle bursts of TRDP_TOPO_ERR and TRDP_NOSUB_ERR better
Since we only fetch one PD at a time, at a fixed interval, erroneous
TRDP PD telegrams can easily get in the way of telegrams that we're actually
interested in. Since even PD telegrams without any subscriber is considered
erroneous in this context this is very likely to happen in an environment where
TRDP is utilized for other things than our part of the -2-3 stack.

To mitigate this, the patch in this commit makes TCNOpen continue to fetch
telegrams until it finds a telegram that doesn't result in TRDP_TOPO_ERR or
TRDP_NOSUB_ERR (as they have been identified as the two most likely errors
that we would get en masse).

If we didn't call tlc_process periodically, but rather as soon as anything is
available on the sockets, we wouldn't have this problem. Switching to an
"event-based" approach is the only viable solution long-term, but short-term
this fix will have to do.